@@ -10,10 +10,6 @@ if defined?(Bundler)
   # Bundler.require(:default, :assets, Rails.env)
 end
 
-CONFIG = YAML.load(ERB.new(File.read(File.expand_path('../application.yml', __FILE__))).result)
-CONFIG.merge! CONFIG.fetch(Rails.env, {})
-CONFIG.symbolize_keys!
-
 module Iftaconferenceapp
   class Application < Rails::Application
     # Settings in config/environments/* take precedence over those specified here.
@@ -52,17 +48,9 @@ module Iftaconferenceapp
     # like if you have constraints or database-specific column types
     # config.active_record.schema_format = :sql
 
-    # Enforce whitelist mode for mass assignment.
-    # This will create an empty whitelist of attributes available for mass-assignment for all models
-    # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
-    # parameters by using an attr_accessible or attr_protected declaration.
-    config.active_record.whitelist_attributes = true
-
-    # Enable the asset pipeline
-    config.assets.enabled = true
-
     # Version of your assets, change this if you want to expire all your assets
     config.assets.version = '1.0'
     config.assets.initialize_on_precompile = false
+    config.active_record.raise_in_transactional_callbacks = true
   end
 end
